+++ branch-master/.artifact/jslint_wrapper_vscode/jslint.mjs
@@ -106,9 +106,9 @@
     formatted_message, free, freeze, from, froms, fsWriteFileWithParents,
     fud_stmt, functionName, function_list, function_stack, functions, get,
     getset, github_repo, globExclude, global, global_dict, global_list,
-    holeList, htmlEscape, id, identifier, import, import_list, import_meta_url,
-    inc, includeList, indent2, index, indexOf, init, initial, isArray,
-    isBlockCoverage, isHole, isNaN, is_equal, is_weird, join, jslint,
+    holeList, htmlEscape, id, identifier, ignoreLine, import, import_list,
+    import_meta_url, inc, includeList, indent2, index, indexOf, init, initial,
+    isArray, isBlockCoverage, isHole, isNaN, is_equal, is_weird, join, jslint,
     jslint_apidoc, jslint_assert, jslint_charset_ascii, jslint_cli,
     jslint_edition, jslint_phase1_split, jslint_phase2_lex, jslint_phase3_parse,
     jslint_phase4_walk, jslint_phase5_whitage, jslint_report, json,
@@ -163,7 +163,7 @@ let jslint_charset_ascii = (
     + "@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_"
     + "`abcdefghijklmnopqrstuvwxyz{|}~\u007f"
 );
-let jslint_edition = "v2025.3.31";
+let jslint_edition = "v2025.12.28";
 let jslint_export;                      // The jslint object to be exported.
 let jslint_fudge = 1;                   // Fudge starting line and starting
                                         // ... column to 1.
@@ -4674,7 +4674,13 @@ function jslint_phase3_parse(state) {
 
             test_cause("free");
             the_paren.free = true;
-            if (the_argument.wrapped === true) {
+            if (
+                the_argument.wrapped === true
+
+// PR-483 - Allow parenthesis after ellipsis inside a function call.
+
+                && the_argument.ellipsis !== true
+            ) {
 
 // test_cause:
 // ["aa((0))", "infix_lparen", "unexpected_a", "(", 3]
@@ -5158,6 +5164,10 @@ function jslint_phase3_parse(state) {
                 the_label.dead = false;
                 the_label.init = true;
                 the_statement = parse_statement();
+
+// Issue #458 - Regression - Warn about variable usage before initialization.
+
+                the_label.dead = true;
                 functionage.statement_prv = the_statement;
                 the_statement.label = the_label;
                 the_statement.statement = true;
@@ -5204,9 +5214,13 @@ function jslint_phase3_parse(state) {
             }
             semicolon();
         }
-        if (the_label !== undefined) {
-            the_label.dead = true;
-        }
+
+// Issue #458 - Regression - Warn about variable usage before initialization.
+
+//        if (the_label !== undefined) {
+//            the_label.dead = true;
+//        }
+
         return the_statement;
     }
 
@@ -6987,8 +7001,16 @@ function jslint_phase3_parse(state) {
                         the_variable.names.push(name);
                         enroll(name, "variable", mode_const);
                     }
-                    name.dead = false;
+
+// Issue #458 - Regression - Warn about variable usage before initialization.
+
+//                    name.dead = false;
+
                     name.init = true;
+
+// test_cause:
+// ["const {aa}=bb;\nconst bb=0;", "lookup", "out_of_scope_a", "bb", 12]
+
                     if (token_nxt.id === "=") {
 
 // test_cause:
@@ -7039,8 +7061,16 @@ function jslint_phase3_parse(state) {
                     advance();
                     the_variable.names.push(name);
                     enroll(name, "variable", mode_const);
-                    name.dead = false;
+
+// Issue #458 - Regression - Warn about variable usage before initialization.
+
+//                    name.dead = false;
+
                     name.init = true;
+
+// test_cause:
+// ["const [aa]=bb;\nconst bb=0;", "lookup", "out_of_scope_a", "bb", 12]
+
                     if (ellipsis) {
                         name.ellipsis = true;
                         break;
@@ -7073,8 +7103,16 @@ function jslint_phase3_parse(state) {
                 enroll(name, "variable", mode_const);
                 if (token_nxt.id === "=" || mode_const) {
                     advance("=");
-                    name.dead = false;
+
+// Issue #458 - Regression - Warn about variable usage before initialization.
+
+//                    name.dead = false;
+
                     name.init = true;
+
+// test_cause:
+// ["const aa=bb;\nconst bb=0;", "lookup", "out_of_scope_a", "bb", 10]
+
                     name.expression = parse_expression(0);
                 }
                 the_variable.names.push(name);
@@ -10882,12 +10920,15 @@ body {
     background: #9d9;
 }
 .coverage .count {
-    color: #666;
+    color: #333;
 }
 .coverage .coverageIgnore {
     background: #ccc;
 }
 .coverage .coverageLow,
+.coverage .ignore {
+    background: #ccc;
+}
 .coverage .uncovered {
     background: #ebb;
 }
@@ -10910,6 +10951,10 @@ body {
 .coverage pre:hover span,
 .coverage tr:hover td {
     background: #7d7;
+}
+.coverage pre:hover span.ignore,
+.coverage tr:hover td.coverageIgnore {
+    background: #ccc;
 }
 .coverage pre:hover span.uncovered,
 .coverage tr:hover td.coverageLow {
@@ -11112,6 +11157,7 @@ body {
             lineList.forEach(function ({
                 count,
                 holeList,
+                ignoreLine,
                 line,
                 startOffset
             }, ii) {
@@ -11159,7 +11205,11 @@ body {
 // true.
 
                             if (isHole) {
-                                lineHtml += " class=\"uncovered\"";
+                                lineHtml += (
+                                    ignoreLine
+                                    ? " class=\"ignore\""
+                                    : " class=\"uncovered\""
+                                );
                             }
                             lineHtml += ">";
                             chunk = "";
@@ -11179,7 +11229,9 @@ body {
 </span>
 <span class="count
                 ${(
-                    count <= 0
+                    (count <= 0 && ignoreLine)
+                    ? "ignore"
+                    : count <= 0
                     ? "uncovered"
                     : ""
                 )}"
@@ -11416,6 +11468,7 @@ function sentinel() {}
         functions,
         url: pathname
     }) {
+        let ignoreBlock = false;
         let lineList;
         let linesCovered;
         let linesTotal;
@@ -11425,14 +11478,23 @@ function sentinel() {}
         source.replace((
             /^.*$/gm
         ), function (line, startOffset) {
+            if (line === "/*coverage-disable*/") {
+                ignoreBlock = true;
+            }
             lineList[lineList.length - 1].endOffset = startOffset - 1;
             lineList.push({
                 count: -1,
                 endOffset: 0,
                 holeList: [],
+                ignoreLine: (
+                    ignoreBlock || line.endsWith("//coverage-ignore-line")
+                ),
                 line,
                 startOffset
             });
+            if (line === "/*coverage-enable*/") {
+                ignoreBlock = false;
+            }
             return "";
         });
         lineList.shift();
@@ -11485,11 +11547,13 @@ function sentinel() {}
             });
         });
         linesTotal = lineList.length;
-        linesCovered = lineList.filter(function ({
-            count
+        linesCovered = 0;
+        lineList.forEach(function ({
+            count,
+            ignoreLine
         }) {
-            return count > 0;
-        }).length;
+            linesCovered += count > 0 || ignoreLine;
+        });
         await moduleFs.promises.mkdir((
             modulePath.dirname(coverageDir + pathname)
         ), {