+++ branch-master/test.mjs
@@ -693,6 +693,19 @@ jstestDescribe((
                 "#!\n/*jslint browser:false, node*/\n\"use strict\";",
                 "/*property aa bb*/"
             ],
+            ellipsis: [
+
+// PR-483 - Allow parenthesis after ellipsis inside a function call.
+
+                (
+                    "let aa = 0;\n"
+                    + "aa(...(\n"
+                    + "    aa()\n"
+                    + "    ? [0]\n"
+                    + "    : [1]\n"
+                    + "));"
+                )
+            ],
             for: [
                 (
                     "/*jslint for*/\n"
@@ -1454,6 +1467,10 @@ jstestDescribe((
                 "/*coverage-ignore-file*/\n"
                 + "switch(0){\n"
                 + "case 0:break;\n"
+                + "case 1:break;//coverage-ignore-line\n"
+                + "/*coverage-disable*/\n"
+                + "case 2:break;\n"
+                + "/*coverage-enable*/\n"
                 + "}\n"
             )
         ], [
@@ -1492,6 +1509,20 @@ jstestDescribe((
                 ]
             });
         });
+    });
+    jstestIt((
+        "test coverage-ignore handling-behavior"
+    ), function () {
+        switch (noop() && noop()) { //coverage-ignore-line
+        case 1: //coverage-ignore-line
+            break; //coverage-ignore-line
+/*coverage-disable*/
+        case 2:
+            break;
+/*coverage-enable*/
+        case undefined:
+            break;
+        }
     });
     jstestIt((
         "test npm handling-behavior"